+++ include/roq/action.hpp
@@ -1,0 +1,18 @@
+/* Copyright (c) 2017-2025, Hans Erik Thrane */
+
+/* !!! THIS FILE HAS BEEN AUTO-GENERATED !!! */
+
+#pragma once
+
+#include <cstdint>
+
+namespace roq {
+
+//! Enumeration of control action
+enum class Action : uint8_t {
+  UNDEFINED = 0,
+  ENABLE,   //!< Enable
+  DISABLE,  //!< Disable
+};
+
+}  // namespace roq