+++ include/roq/control.hpp
@@ -1,0 +1,77 @@
+/* Copyright (c) 2017-2025, Hans Erik Thrane */
+
+/* !!! THIS FILE HAS BEEN AUTO-GENERATED !!! */
+
+#pragma once
+
+#include "roq/compat.hpp"
+
+#include <fmt/core.h>
+
+#include <magic_enum/magic_enum_format.hpp>
+
+#include "roq/action.hpp"
+#include "roq/event.hpp"
+#include "roq/name.hpp"
+#include "roq/trace.hpp"
+
+namespace roq {
+
+//! Control service state
+struct ROQ_PUBLIC Control final {
+  roq::Action action = {};  //!< Action
+};
+
+template <>
+inline constexpr std::string_view get_name<Control>() {
+  using namespace std::literals;
+  return "control"sv;
+}
+
+}  // namespace roq
+
+template <>
+struct fmt::formatter<roq::Control> {
+  constexpr auto parse(format_parse_context &context) { return std::begin(context); }
+  auto format(roq::Control const &value, format_context &context) const {
+    using namespace std::literals;
+    return fmt::format_to(
+        context.out(),
+        R"({{)"
+        R"(action={})"
+        R"(}})"sv,
+        value.action);
+  }
+};
+
+template <>
+struct fmt::formatter<roq::Event<roq::Control>> {
+  constexpr auto parse(format_parse_context &context) { return std::begin(context); }
+  auto format(roq::Event<roq::Control> const &event, format_context &context) const {
+    using namespace std::literals;
+    return fmt::format_to(
+        context.out(),
+        R"({{)"
+        R"(control={}, )"
+        R"(message_info={})"
+        R"(}})"sv,
+        event.value,
+        event.message_info);
+  }
+};
+
+template <>
+struct fmt::formatter<roq::Trace<roq::Control>> {
+  constexpr auto parse(format_parse_context &context) { return std::begin(context); }
+  auto format(roq::Trace<roq::Control> const &event, format_context &context) const {
+    using namespace std::literals;
+    return fmt::format_to(
+        context.out(),
+        R"({{)"
+        R"(control={}, )"
+        R"(trace_info={})"
+        R"(}})"sv,
+        event.value,
+        event.trace_info);
+  }
+};